[refactor] ReadBuffer 중복 제거(클래스 상속으로 변경) 및 패키지 이동, 클래스명 변경
